Common types module (#259)
* Add common types module

* Make course types common

* Add build-common steps

* Fix Docker builds
