+++ client/package.json
@@ -4,6 +4,9 @@
   "author": "The Aalto Grades Developers",
   "license": "Expat (MIT)",
   "private": true,
+  "workspaces": [
+    "../common"
+  ],
   "dependencies": {
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
@@ -21,6 +24,7 @@
     "@types/react-dom": "^18.2.4",
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/parser": "^5.59.8",
+    "aalto-grades-common": "*",
     "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -31,7 +35,7 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "npm --prefix ../common run build && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
