+++ client/src/components/auth/Signup.tsx
@@ -8,7 +8,7 @@ import { useTheme } from '@mui/material/styles';
 import SignupForm from './SignupForm';
 import userService from '../../services/user';
 import useAuth from '../../hooks/useAuth';
-import { SystemRole } from '../../types/general';
+import { SystemRole } from 'aalto-grades-common/types/general';
 
 const Signup = () => {
