+++ client/src/services/courses.tsx
@@ -3,27 +3,33 @@
 // SPDX-License-Identifier: MIT
 
 import axios from './axios';
+import { CourseData } from 'aalto-grades-common/types/course';
 
-const getCoursesOfUser: any = async (userId) => {
+async function getCoursesOfUser(userId: number): Promise<{
+  courses: {
+    current: Array<CourseData>,
+    previous: Array<CourseData>
+  }
+}> {
   const response = await axios.get(`/v1/user/${userId}/courses`);
   console.log(response.data);
   return response.data.data;
-};
+}
 
-const getCourse: any = async (courseId) => {
+async function getCourse(courseId: number): Promise<{ course: CourseData }> {
   const response = await axios.get(`/v1/courses/${courseId}`);
   return response.data.data;
-};
+}
 
-const getAllCourses: any = async () => {
+async function getAllCourses(): Promise<{ courses: Array<CourseData> }> {
   const response = await axios.get('/v1/courses');
   return response.data.data;
-};
+}
 
 // .data added here too, not tested though
-const addCourse: any = async (course) => {
+async function addCourse(course: CourseData): Promise<{ course: { id: number } }> {
   const response = await axios.post('/v1/courses', course);
   return response.data.data;
-};
+}
 
 export default { getCoursesOfUser, getCourse, getAllCourses, addCourse };
