+++ client/src/tests/CourseView.test.tsx
@@ -13,7 +13,7 @@ import instancesService from '../services/instances';
 import AuthContext from '../context/authProvider';
 import mockCourses from '../mock-data/mockCourses';
 import mockInstances from '../mock-data/mockInstancesWithStringDates';
-import { SystemRole } from '../types/general';
+import { SystemRole } from 'aalto-grades-common/types/general';
 
 jest.mock('../services/courses');
 jest.mock('../services/instances');
@@ -29,8 +29,8 @@ describe('Tests for CourseView component', () => {
 
 
     const mockResponseCourse = { course: mockCourses[0] };
-    coursesService.getCourse.mockRejectedValue('Network error');
-    coursesService.getCourse.mockResolvedValue(mockResponseCourse);
+    (coursesService.getCourse as jest.Mock).mockRejectedValue('Network error');
+    (coursesService.getCourse as jest.Mock).mockResolvedValue(mockResponseCourse);
 
     return render(
       <MemoryRouter initialEntries={['/course-view/1']}>
