--- client/src/types/general.ts
@@ -1,15 +1,0 @@
-// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
-//
-// SPDX-License-Identifier: MIT
-
-/* eslint-disable no-unused-vars */
-export enum SystemRole {
-  User = 'USER',
-  Admin = 'ADMIN'
-}
-
-export enum CourseInstanceRole {
-  Student = 'STUDENT',
-  Teacher = 'TEACHER',
-  TeacherInCharge = 'TEACHER_IN_CHARGE',
-}
