+++ common-services.yaml
@@ -10,7 +10,8 @@ version: "3.7"
 services:
   frontend:
     build:
-      context: ./client
+      context: .
+      dockerfile: ./client/Dockerfile
       args:
         CLIENT_SKIP_NPM_CI: ${CLIENT_SKIP_NPM_CI}
     extends:
@@ -27,7 +28,8 @@ services:
 
   backend:
     build:
-      context: ./server
+      context: .
+      dockerfile: ./server/Dockerfile
       args:
         SERVER_SKIP_NPM_CI: ${SERVER_SKIP_NPM_CI}
     extends:
