+++ common/.gitignore
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+#
+# SPDX-License-Identifier: MIT
+
+*.js
