+++ common/types/course.ts
