+++ common/types/general.ts
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: 2023 The Aalto Grades Developers
+//
+// SPDX-License-Identifier: MIT
+
+export enum SystemRole {
+  User = 'USER',
+  Admin = 'ADMIN'
+}
